+++ src/button/meson.build
@@ -1,0 +1,3 @@
+labwc_sources += files(
+  'button-xbm.c',
+)
