--- src/xbm/meson.build
@@ -1,5 +1,0 @@
-labwc_sources += files(
-  'parse.c',
-  'tokenize.c',
-  'xbm.c',
-)
